requirements.txt
attrs==22.2.0
charset-normalizer==3.0.1
comet-ml==3.31.22
configobj==5.0.6
dulwich==0.21.0
everett==3.1.0
idna==3.4
importlib-metadata==6.1.0
importlib-resources==5.10.2
jsonschema==4.17.3
MarkupSafe==2.1.2
nflows==0.14
numpy==1.23.2
oauthlib==3.2.2
packaging==21.3
Pillow==9.2.0
pkgutil_resolve_name==1.3.10
pyro-api==0.1.2
pyro-ppl==1.8.4
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
requests==2.28.2
requests-toolbelt==0.10.1
semantic-version==2.10.0
sentry-sdk==1.13.0
simplejson==3.18.1
six==1.16.0
torch==1.9.0+cu111
torchaudio==0.9.0
torchvision==0.10.0+cu111
typing_extensions==4.3.0
urllib3==1.26.14
websocket-client==1.3.3
Werkzeug==2.2.3
wrapt==1.14.1
wurlitzer==3.0.3
zipp==3.11.0
